Rearrange templates
